create writer and cleanup
