+++ tools/src/main/scala/com/nvidia/spark/rapids/tool/qualification/Qualification.scala
@@ -18,25 +18,24 @@ package com.nvidia.spark.rapids.tool.qualification
 
 import scala.collection.mutable.ArrayBuffer
 
-import com.nvidia.spark.rapids.tool.{EventLogInfo, ToolTextFileWriter}
+import com.nvidia.spark.rapids.tool.EventLogInfo
 
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.rapids.tool.qualification._
 
 /**
- * Ranks the applications for GPU acceleration.
+ * Scores the applications for GPU acceleration and outputs the
+ * reports.
  */
-class Qualification(outputDir: String) extends Logging {
-
-  val finalOutputDir = s"$outputDir/rapids_4_spark_qualification_output"
-  val logFileName = "rapids_4_spark_qualification_output"
+object Qualification extends Logging {
 
   def qualifyApps(
       allPaths: Seq[EventLogInfo],
-      numRows: Int): ArrayBuffer[QualificationSummaryInfo] = {
+      numRows: Int,
+      outputDir: String): ArrayBuffer[QualificationSummaryInfo] = {
     val allAppsSum: ArrayBuffer[QualificationSummaryInfo] = ArrayBuffer[QualificationSummaryInfo]()
     allPaths.foreach { path =>
-      val (app, _) = QualAppInfo.createApp(path, numRows)
+      val app = QualAppInfo.createApp(path, numRows)
       if (!app.isDefined) {
         logWarning("No Applications found that contain SQL!")
       } else {
@@ -49,77 +48,9 @@ class Qualification(outputDir: String) extends Logging {
       }
     }
     val sorted = allAppsSum.sortBy(sum => (-sum.score, -sum.sqlDataFrameDuration, -sum.appDuration))
-    val csvFileWriter = new ToolTextFileWriter(finalOutputDir, s"${logFileName}.csv")
-    try {
-      writeCSVHeader(csvFileWriter)
-      sorted.foreach { appSum =>
-        writeCSV(csvFileWriter, appSum)
-      }
-    } finally {
-      csvFileWriter.close()
-
-    }
-    val textFileWriter = new ToolTextFileWriter(finalOutputDir, s"${logFileName}.log")
-    try {
-      writeTextSummary(textFileWriter, sorted)
-    } finally {
-      textFileWriter.close()
-    }
+    val qWriter = new QualOutputWriter(outputDir)
+    qWriter.writeCSV(sorted)
+    qWriter.writeReport(sorted)
     sorted
-  }
-
-  val problemDurStr = "SQL Duration For Problematic"
-  val headers = Array("App ID", "App Duration", "SQL Dataframe Duration", problemDurStr)
-
-  private def getTextSpacing(sums: Seq[QualificationSummaryInfo]): (Int, Int)= {
-    val sizePadLongs = problemDurStr.size
-    val sizes = sums.map(_.appId.size)
-    val appIdMaxSize = if (sizes.size > 0) sizes.max else 0
-    (appIdMaxSize, sizePadLongs)
-  }
-
-  def headerCSV: String = {
-    "App Name,App ID,Score,Potential Problems,SQL Dataframe Duration," +
-      "App Duration,Executor CPU Time Percent,App Duration Estimated," +
-      "SQL Duration with Potential Problems,SQL Ids with Failures\n"
-  }
-
-  def writeCSVHeader(writer: ToolTextFileWriter): Unit = {
-    writer.write(headerCSV)
-  }
-
-  def writeCSV(writer: ToolTextFileWriter, sumInfo: QualificationSummaryInfo): Unit = {
-    writer.write(sumInfo.toCSV + "\n")
-  }
-
-  def writeTextSummary(writer: ToolTextFileWriter,
-      sums: Seq[QualificationSummaryInfo]): Unit = {
-    val (appIdMaxSize, sizePadLongs) = getTextSpacing(sums)
-    val entireHeader = new StringBuffer
-    entireHeader.append("|")
-    val appIdSpaces = " " * (appIdMaxSize - headers(0).size)
-    entireHeader.append(s"$appIdSpaces${headers(0)}|")
-    entireHeader.append(s"${" " * (sizePadLongs - headers(1).size)}${headers(1)}|")
-    entireHeader.append(s"${" " * (sizePadLongs - headers(2).size)}${headers(2)}|")
-    entireHeader.append(s"${" " * (sizePadLongs - headers(3).size)}${headers(3)}|")
-    entireHeader.append("\n")
-    val sep = "=" * (appIdMaxSize + (sizePadLongs * 3) + 5)
-    writer.write(s"$sep\n")
-    writer.write(entireHeader.toString)
-    writer.write(s"$sep\n")
-
-    sums.foreach { sumInfo =>
-      val appId = sumInfo.appId
-      val appPad = " " * (appIdMaxSize - appId.size)
-      val appDur = sumInfo.appDuration.toString
-      val sqlDur = sumInfo.sqlDataFrameDuration.toString
-      val sqlProbDur = sumInfo.sqlDurationForProblematic.toString
-      val appDurPad = " " * (sizePadLongs - appDur.size)
-      val sqlDurPad = " " * (sizePadLongs - sqlDur.size)
-      val sqlProbDurPad = " " * (sizePadLongs - sqlProbDur.size)
-      val wStr = s"|$appPad$appId|$appDurPad$appDur|$sqlDurPad$sqlDur|$sqlProbDurPad$sqlProbDur|"
-      writer.write(wStr + "\n")
-    }
-    writer.write(s"$sep\n")
   }
 }
