+++ tools/src/main/scala/org/apache/spark/sql/rapids/tool/qualification/QualAppInfo.scala
@@ -16,25 +16,19 @@
 
 package org.apache.spark.sql.rapids.tool.qualification
 
-import scala.collection.mutable.{ArrayBuffer, HashMap}
-import scala.io.{Codec, Source}
+import scala.collection.mutable.{ArrayBuffer, HashMap, HashSet}
 
-import com.nvidia.spark.rapids.tool.{DatabricksEventLog, DatabricksRollingEventLogFilesFileReader, EventLogInfo, EventLogPathProcessor}
+import com.nvidia.spark.rapids.tool.EventLogInfo
 import com.nvidia.spark.rapids.tool.profiling._
-import org.apache.hadoop.conf.Configuration
-import org.json4s.jackson.JsonMethods.parse
 
-import org.apache.spark.deploy.history.EventLogFileReader
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.execution.SparkPlanInfo
 import org.apache.spark.sql.execution.ui.SparkPlanGraph
-import org.apache.spark.sql.rapids.tool.{AppBase, ToolUtils}
-import org.apache.spark.util.{JsonProtocol, Utils}
+import org.apache.spark.sql.rapids.tool.AppBase
 
 class QualAppInfo(
     numOutputRows: Int,
-    eventLogInfo: EventLogInfo,
-    val index: Int) extends AppBase(numOutputRows, eventLogInfo) with Logging {
+    eventLogInfo: EventLogInfo) extends AppBase(numOutputRows, eventLogInfo) with Logging {
 
   var appId: String = ""
   var isPluginEnabled = false
@@ -57,8 +51,7 @@ class QualAppInfo(
   val problematicSQL: ArrayBuffer[ProblematicSQLCase] = ArrayBuffer[ProblematicSQLCase]()
 
   // SQL containing any Dataset operation
-  // TODO - don't really need the DatasetSQLCase
-  val sqlIDToDataSetCase: HashMap[Long, DatasetSQLCase] = HashMap[Long, DatasetSQLCase]()
+  val sqlIDToDataSetCase: HashSet[Long] = HashSet[Long]()
 
   processEvents()
 
@@ -83,6 +76,7 @@ class QualAppInfo(
     ProfileUtils.OptionLongMinusLong(estimatedResult, startTime)
   }
 
+  // get percent to 2 decimal places
   private def calculatePercent(first: Long, total: Long): Double = {
     val firstDec = BigDecimal.decimal(first)
     val totalDec = BigDecimal.decimal(total)
@@ -132,7 +126,6 @@ class QualAppInfo(
 
   def aggregateStats(): Option[QualificationSummaryInfo] = {
     appInfo.map { info =>
-      // TODO - do we want 0 here or something else?
       val appDuration = calculateAppDuration(info.startTime).getOrElse(0L)
       val sqlDataframeDur = calculateSqlDataframDuration
       val score = calculateScore(sqlDataframeDur, appDuration)
@@ -149,80 +142,18 @@ class QualAppInfo(
     }
   }
 
-  /**
-   * Functions to process all the events
-   */
-  // TODO - commonize
-  def processEvents(): Unit = {
-    val eventlog = eventLogInfo.eventLog
-
-    logInfo("Parsing Event Log: " + eventlog.toString)
-
-    // at this point all paths should be valid event logs or event log dirs
-    // TODO - reuse Configuration
-    val fs = eventlog.getFileSystem(new Configuration)
-    var totalNumEvents = 0
-    val eventsProcessor = new QualEventProcessor()
-    val readerOpt = eventLogInfo match {
-      case dblog: DatabricksEventLog =>
-        Some(new DatabricksRollingEventLogFilesFileReader(fs, eventlog))
-      case apachelog => EventLogFileReader(fs, eventlog)
-    }
-
-    if (readerOpt.isDefined) {
-      val reader = readerOpt.get
-      val logFiles = reader.listEventLogFiles
-      logFiles.foreach { file =>
-        Utils.tryWithResource(ToolUtils.openEventLogInternal(file.getPath, fs)) { in =>
-          val lines = Source.fromInputStream(in)(Codec.UTF8).getLines().toList
-          totalNumEvents += lines.size
-          lines.foreach { line =>
-            try {
-              val event = JsonProtocol.sparkEventFromJson(parse(line))
-              eventsProcessor.processAnyEvent(this, event)
-            }
-            catch {
-              case e: ClassNotFoundException =>
-                logWarning(s"ClassNotFoundException: ${e.getMessage}")
-            }
-          }
-        }
-      }
-    } else {
-      logError(s"Error getting reader for ${eventlog.getName}")
-    }
-    logInfo("Total number of events parsed: " + totalNumEvents)
-  }
-
   def processSQLPlan(sqlID: Long, planInfo: SparkPlanInfo): Unit = {
     val planGraph = SparkPlanGraph(planInfo)
     val allnodes = planGraph.allNodes
     for (node <- allnodes) {
       if (isDataSetPlan(node.desc)) {
-        sqlIDToDataSetCase += (sqlID -> DatasetSQLCase(sqlID))
+        sqlIDToDataSetCase += sqlID
       }
       findPotentialIssues(node.desc).foreach { issues =>
         problematicSQL += ProblematicSQLCase(sqlID, issues)
       }
     }
   }
-
-  // TODO - commonize
-  def isDataSetPlan(desc: String): Boolean = {
-    desc match {
-      case l if l.matches(".*\\$Lambda\\$.*") => true
-      case a if a.endsWith(".apply") => true
-      case _ => false
-    }
-  }
-
-  def findPotentialIssues(desc: String): Option[String] =  {
-    desc match {
-      case u if u.matches(".*UDF.*") => Some("UDF")
-      case _ => None
-    }
-  }
-
 }
 
 case class QualApplicationInfo(
@@ -254,56 +185,28 @@ case class QualificationSummaryInfo(
     executorCpuTimePercent: Double,
     endDurationEstimated: Boolean,
     sqlDurationForProblematic: Long,
-    failedSQLIds: String) {
-
-  private def stringIfempty(str: String): String = {
-    if (str.isEmpty) {
-      "\"\""
-    } else {
-      str
-    }
-  }
-
-  def toCSV: String = {
-    val probStr = stringIfempty(potentialProblems)
-    val appIdStr = stringIfempty(appId)
-    val appNameStr = stringIfempty(appName)
-    val failedIds = stringIfempty(failedSQLIds)
-
-    s"$appNameStr,$appIdStr,$score,$probStr,$sqlDataFrameDuration," +
-      s"$appDuration,$executorCpuTimePercent," +
-      s"$endDurationEstimated,$sqlDurationForProblematic,$failedIds"
-  }
-}
+    failedSQLIds: String)
 
 object QualAppInfo extends Logging {
   def createApp(
       path: EventLogInfo,
-      numRows: Int,
-      startIndex: Int = 1): (Option[QualAppInfo], Int) = {
-    var index: Int = startIndex
-    var errorCode = 0
+      numRows: Int): Option[QualAppInfo] = {
     val app = try {
-        // This apps only contains 1 app in each loop.
-        val app = new QualAppInfo(numRows, path, index)
-        logInfo(s"==============  ${app.appId} (index=${app.index})  ==============")
-        index += 1
+        val app = new QualAppInfo(numRows, path)
+        logInfo(s"==============  ${app.appId} ============== ")
         Some(app)
       } catch {
         case json: com.fasterxml.jackson.core.JsonParseException =>
           logWarning(s"Error parsing JSON: $path")
-          errorCode = 1
           None
         case il: IllegalArgumentException =>
           logWarning(s"Error parsing file: $path", il)
-          errorCode = 2
           None
         case e: Exception =>
           // catch all exceptions and skip that file
           logWarning(s"Got unexpected exception processing file: $path", e)
-          errorCode = 3
           None
       }
-    (app, errorCode)
+    app
   }
 }
